+++ hookQuickBuild/programSet.cpp
@@ -5,11 +5,11 @@
 std::string LPCWSTRTostring(LPCWSTR pWCStrKey)
 {
     //第一次调用确认转换后单字节字符串的长度，用于开辟空间
-    int pSize = WideCharToMultiByte(CP_OEMCP, 0, pWCStrKey, wcslen(pWCStrKey),
+    int pSize = WideCharToMultiByte(CP_OEMCP, 0, pWCStrKey, (int)wcslen(pWCStrKey),
         NULL, 0, NULL, NULL);
     char* pCStrKey = new char[pSize + 1];
     //第二次调用将双字节字符串转换成单字节字符串
-    WideCharToMultiByte(CP_OEMCP, 0, pWCStrKey, wcslen(pWCStrKey),
+    WideCharToMultiByte(CP_OEMCP, 0, pWCStrKey, (int)wcslen(pWCStrKey),
         pCStrKey, pSize, NULL, NULL);
     
     pCStrKey[pSize] = '\0';
